Move internal into internal/buffer
