+++ node/_buffer.js
@@ -9,744 +9,53 @@ import {
 } from "./_errors.ts";
 import { isAnyArrayBuffer, isArrayBufferView } from "./_util/_util_types.ts";
 import { normalizeEncoding } from "./internal/util.ts";
-
-// Temporary buffers to convert numbers.
-const float32Array = new Float32Array(1);
-const uInt8Float32Array = new Uint8Array(float32Array.buffer);
-const float64Array = new Float64Array(1);
-const uInt8Float64Array = new Uint8Array(float64Array.buffer);
-
-// Check endianness.
-float32Array[0] = -1; // 0xBF800000
-// Either it is [0, 0, 128, 191] or [191, 128, 0, 0]. It is not possible to
-// check this with `os.endianness()` because that is determined at compile time.
-const bigEndian = uInt8Float32Array[3] === 0;
-
-function readUInt48LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 5];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 6);
-  }
-
-  return first +
-    buf[++offset] * 2 ** 8 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 24 +
-    (buf[++offset] + last * 2 ** 8) * 2 ** 32;
-}
-
-function readUInt40LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 4];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 5);
-  }
-
-  return first +
-    buf[++offset] * 2 ** 8 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 24 +
-    last * 2 ** 32;
-}
-
-function readUInt24LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 2];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 3);
-  }
-
-  return first + buf[++offset] * 2 ** 8 + last * 2 ** 16;
-}
-
-function readUInt48BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 5];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 6);
-  }
-
-  return (first * 2 ** 8 + buf[++offset]) * 2 ** 32 +
-    buf[++offset] * 2 ** 24 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 8 +
-    last;
-}
-
-function readUInt40BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 4];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 5);
-  }
-
-  return first * 2 ** 32 +
-    buf[++offset] * 2 ** 24 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 8 +
-    last;
-}
-
-function readUInt24BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 2];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 3);
-  }
-
-  return first * 2 ** 16 + buf[++offset] * 2 ** 8 + last;
-}
-
-function readUInt16BE(offset = 0) {
-  validateNumber(offset, "offset");
-  const first = this[offset];
-  const last = this[offset + 1];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, this.length - 2);
-  }
-
-  return first * 2 ** 8 + last;
-}
-
-function readUInt32BE(offset = 0) {
-  validateNumber(offset, "offset");
-  const first = this[offset];
-  const last = this[offset + 3];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, this.length - 4);
-  }
-
-  return first * 2 ** 24 +
-    this[++offset] * 2 ** 16 +
-    this[++offset] * 2 ** 8 +
-    last;
-}
-
-function readDoubleBackwards(buffer, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buffer[offset];
-  const last = buffer[offset + 7];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buffer.length - 8);
-  }
-
-  uInt8Float64Array[7] = first;
-  uInt8Float64Array[6] = buffer[++offset];
-  uInt8Float64Array[5] = buffer[++offset];
-  uInt8Float64Array[4] = buffer[++offset];
-  uInt8Float64Array[3] = buffer[++offset];
-  uInt8Float64Array[2] = buffer[++offset];
-  uInt8Float64Array[1] = buffer[++offset];
-  uInt8Float64Array[0] = last;
-  return float64Array[0];
-}
-
-function readDoubleForwards(buffer, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buffer[offset];
-  const last = buffer[offset + 7];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buffer.length - 8);
-  }
-
-  uInt8Float64Array[0] = first;
-  uInt8Float64Array[1] = buffer[++offset];
-  uInt8Float64Array[2] = buffer[++offset];
-  uInt8Float64Array[3] = buffer[++offset];
-  uInt8Float64Array[4] = buffer[++offset];
-  uInt8Float64Array[5] = buffer[++offset];
-  uInt8Float64Array[6] = buffer[++offset];
-  uInt8Float64Array[7] = last;
-  return float64Array[0];
-}
-
-function writeDoubleForwards(buffer, val, offset = 0) {
-  val = +val;
-  checkBounds(buffer, offset, 7);
-
-  float64Array[0] = val;
-  buffer[offset++] = uInt8Float64Array[0];
-  buffer[offset++] = uInt8Float64Array[1];
-  buffer[offset++] = uInt8Float64Array[2];
-  buffer[offset++] = uInt8Float64Array[3];
-  buffer[offset++] = uInt8Float64Array[4];
-  buffer[offset++] = uInt8Float64Array[5];
-  buffer[offset++] = uInt8Float64Array[6];
-  buffer[offset++] = uInt8Float64Array[7];
-  return offset;
-}
-
-function writeDoubleBackwards(buffer, val, offset = 0) {
-  val = +val;
-  checkBounds(buffer, offset, 7);
-
-  float64Array[0] = val;
-  buffer[offset++] = uInt8Float64Array[7];
-  buffer[offset++] = uInt8Float64Array[6];
-  buffer[offset++] = uInt8Float64Array[5];
-  buffer[offset++] = uInt8Float64Array[4];
-  buffer[offset++] = uInt8Float64Array[3];
-  buffer[offset++] = uInt8Float64Array[2];
-  buffer[offset++] = uInt8Float64Array[1];
-  buffer[offset++] = uInt8Float64Array[0];
-  return offset;
-}
-
-function readFloatBackwards(buffer, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buffer[offset];
-  const last = buffer[offset + 3];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buffer.length - 4);
-  }
-
-  uInt8Float32Array[3] = first;
-  uInt8Float32Array[2] = buffer[++offset];
-  uInt8Float32Array[1] = buffer[++offset];
-  uInt8Float32Array[0] = last;
-  return float32Array[0];
-}
-
-function readFloatForwards(buffer, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buffer[offset];
-  const last = buffer[offset + 3];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buffer.length - 4);
-  }
-
-  uInt8Float32Array[0] = first;
-  uInt8Float32Array[1] = buffer[++offset];
-  uInt8Float32Array[2] = buffer[++offset];
-  uInt8Float32Array[3] = last;
-  return float32Array[0];
-}
-
-function writeFloatForwards(buffer, val, offset = 0) {
-  val = +val;
-  checkBounds(buffer, offset, 3);
-
-  float32Array[0] = val;
-  buffer[offset++] = uInt8Float32Array[0];
-  buffer[offset++] = uInt8Float32Array[1];
-  buffer[offset++] = uInt8Float32Array[2];
-  buffer[offset++] = uInt8Float32Array[3];
-  return offset;
-}
-
-function writeFloatBackwards(buffer, val, offset = 0) {
-  val = +val;
-  checkBounds(buffer, offset, 3);
-
-  float32Array[0] = val;
-  buffer[offset++] = uInt8Float32Array[3];
-  buffer[offset++] = uInt8Float32Array[2];
-  buffer[offset++] = uInt8Float32Array[1];
-  buffer[offset++] = uInt8Float32Array[0];
-  return offset;
-}
-
-function readInt24LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 2];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 3);
-  }
-
-  const val = first + buf[++offset] * 2 ** 8 + last * 2 ** 16;
-  return val | (val & 2 ** 23) * 0x1fe;
-}
-
-function readInt40LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 4];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 5);
-  }
-
-  return (last | (last & 2 ** 7) * 0x1fffffe) * 2 ** 32 +
-    first +
-    buf[++offset] * 2 ** 8 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 24;
-}
-
-function readInt48LE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 5];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 6);
-  }
-
-  const val = buf[offset + 4] + last * 2 ** 8;
-  return (val | (val & 2 ** 15) * 0x1fffe) * 2 ** 32 +
-    first +
-    buf[++offset] * 2 ** 8 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 24;
-}
-
-function readInt24BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 2];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 3);
-  }
-
-  const val = first * 2 ** 16 + buf[++offset] * 2 ** 8 + last;
-  return val | (val & 2 ** 23) * 0x1fe;
-}
-
-function readInt48BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 5];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 6);
-  }
-
-  const val = buf[++offset] + first * 2 ** 8;
-  return (val | (val & 2 ** 15) * 0x1fffe) * 2 ** 32 +
-    buf[++offset] * 2 ** 24 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 8 +
-    last;
-}
-
-function readInt40BE(buf, offset = 0) {
-  validateNumber(offset, "offset");
-  const first = buf[offset];
-  const last = buf[offset + 4];
-  if (first === undefined || last === undefined) {
-    boundsError(offset, buf.length - 5);
-  }
-
-  return (first | (first & 2 ** 7) * 0x1fffffe) * 2 ** 32 +
-    buf[++offset] * 2 ** 24 +
-    buf[++offset] * 2 ** 16 +
-    buf[++offset] * 2 ** 8 +
-    last;
-}
-
-function base64UrlToBytes(str) {
-  return new Uint8Array(
-    [...atob(str.replace(/-/g, "+").replace(/_/g, "/"))].map((val) =>
-      val.charCodeAt(0)
-    ),
-  );
-}
-
-function byteLengthUtf8(str) {
-  return new TextEncoder().encode(str).length;
-}
-
-function base64ByteLength(str, bytes) {
-  // Handle padding
-  if (str.charCodeAt(bytes - 1) === 0x3D) {
-    bytes--;
-  }
-  if (bytes > 1 && str.charCodeAt(bytes - 1) === 0x3D) {
-    bytes--;
-  }
-
-  // Base64 ratio: 3/4
-  return (bytes * 3) >>> 2;
-}
-
-const encodingsMap = {};
-
-const encodingOps = {
-  utf8: {
-    encoding: "utf8",
-    encodingVal: encodingsMap.utf8,
-    byteLength: byteLengthUtf8,
-    write: (buf, string, offset, len) => buf.utf8Write(string, offset, len),
-    slice: (buf, start, end) => buf.utf8Slice(start, end),
-  },
-  ucs2: {
-    encoding: "ucs2",
-    encodingVal: encodingsMap.utf16le,
-    byteLength: (string) => string.length * 2,
-    write: (buf, string, offset, len) => buf.ucs2Write(string, offset, len),
-    slice: (buf, start, end) => buf.ucs2Slice(start, end),
-  },
-  utf16le: {
-    encoding: "utf16le",
-    encodingVal: encodingsMap.utf16le,
-    byteLength: (string) => string.length * 2,
-    write: (buf, string, offset, len) => buf.ucs2Write(string, offset, len),
-    slice: (buf, start, end) => buf.ucs2Slice(start, end),
-  },
-  latin1: {
-    encoding: "latin1",
-    encodingVal: encodingsMap.latin1,
-    byteLength: (string) => string.length,
-    write: (buf, string, offset, len) => buf.latin1Write(string, offset, len),
-    slice: (buf, start, end) => buf.latin1Slice(start, end),
-  },
-  ascii: {
-    encoding: "ascii",
-    encodingVal: encodingsMap.ascii,
-    byteLength: (string) => string.length,
-    write: (buf, string, offset, len) => buf.asciiWrite(string, offset, len),
-    slice: (buf, start, end) => buf.asciiSlice(start, end),
-  },
-  base64: {
-    encoding: "base64",
-    encodingVal: encodingsMap.base64,
-    byteLength: (string) => base64ByteLength(string, string.length),
-    write: (buf, string, offset, len) => buf.base64Write(string, offset, len),
-    slice: (buf, start, end) => buf.base64Slice(start, end),
-  },
-  base64url: {
-    encoding: "base64url",
-    encodingVal: encodingsMap.base64url,
-    byteLength: (string) => base64ByteLength(string, string.length),
-    write: (buf, string, offset, len) =>
-      buf.base64urlWrite(string, offset, len),
-    slice: (buf, start, end) => buf.base64urlSlice(start, end),
-  },
-  hex: {
-    encoding: "hex",
-    encodingVal: encodingsMap.hex,
-    byteLength: (string) => string.length >>> 1,
-    write: (buf, string, offset, len) => buf.hexWrite(string, offset, len),
-    slice: (buf, start, end) => buf.hexSlice(start, end),
-  },
-};
-
-function getEncodingOps(encoding) {
-  encoding = String(encoding).toLowerCase();
-  switch (encoding.length) {
-    case 4:
-      if (encoding === "utf8") return encodingOps.utf8;
-      if (encoding === "ucs2") return encodingOps.ucs2;
-      break;
-    case 5:
-      if (encoding === "utf-8") return encodingOps.utf8;
-      if (encoding === "ascii") return encodingOps.ascii;
-      if (encoding === "ucs-2") return encodingOps.ucs2;
-      break;
-    case 7:
-      if (encoding === "utf16le") {
-        return encodingOps.utf16le;
-      }
-      break;
-    case 8:
-      if (encoding === "utf-16le") {
-        return encodingOps.utf16le;
-      }
-      break;
-    case 6:
-      if (encoding === "latin1" || encoding === "binary") {
-        return encodingOps.latin1;
-      }
-      if (encoding === "base64") return encodingOps.base64;
-    case 3:
-      if (encoding === "hex") {
-        return encodingOps.hex;
-      }
-      break;
-    case 9:
-      if (encoding === "base64url") {
-        return encodingOps.base64url;
-      }
-      break;
-  }
-}
-
-function _copyActual(source, target, targetStart, sourceStart, sourceEnd) {
-  if (sourceEnd - sourceStart > target.length - targetStart) {
-    sourceEnd = sourceStart + target.length - targetStart;
-  }
-
-  let nb = sourceEnd - sourceStart;
-  const sourceLen = source.length - sourceStart;
-  if (nb > sourceLen) {
-    nb = sourceLen;
-  }
-
-  if (sourceStart !== 0 || sourceEnd < source.length) {
-    source = new Uint8Array(source.buffer, source.byteOffset + sourceStart, nb);
-  }
-
-  target.set(source, targetStart);
-
-  return nb;
-}
-
-function boundsError(value, length, type) {
-  if (Math.floor(value) !== value) {
-    validateNumber(value, type);
-    throw new ERR_OUT_OF_RANGE(type || "offset", "an integer", value);
-  }
-
-  if (length < 0) {
-    throw new ERR_BUFFER_OUT_OF_BOUNDS();
-  }
-
-  throw new ERR_OUT_OF_RANGE(
-    type || "offset",
-    `>= ${type ? 1 : 0} and <= ${length}`,
-    value,
-  );
-}
-
-function isUint8Array(element) {
-  return element instanceof Uint8Array;
-}
-
-function validateNumber(value, name) {
-  if (typeof value !== "number") {
-    throw new ERR_INVALID_ARG_TYPE(name, "number", value);
-  }
-}
-
-function checkBounds(buf, offset, byteLength) {
-  validateNumber(offset, "offset");
-  if (buf[offset] === undefined || buf[offset + byteLength] === undefined) {
-    boundsError(offset, buf.length - (byteLength + 1));
-  }
-}
-
-function checkInt(value, min, max, buf, offset, byteLength) {
-  if (value > max || value < min) {
-    const n = typeof min === "bigint" ? "n" : "";
-    let range;
-    if (byteLength > 3) {
-      if (min === 0 || min === 0n) {
-        range = `>= 0${n} and < 2${n} ** ${(byteLength + 1) * 8}${n}`;
-      } else {
-        range = `>= -(2${n} ** ${(byteLength + 1) * 8 - 1}${n}) and ` +
-          `< 2${n} ** ${(byteLength + 1) * 8 - 1}${n}`;
-      }
-    } else {
-      range = `>= ${min}${n} and <= ${max}${n}`;
-    }
-    throw new ERR_OUT_OF_RANGE("value", range, value);
-  }
-  checkBounds(buf, offset, byteLength);
-}
-
-function toInteger(n, defaultVal) {
-  n = +n;
-  if (
-    !Number.isNaN(n) &&
-    n >= Number.MIN_SAFE_INTEGER &&
-    n <= Number.MAX_SAFE_INTEGER
-  ) {
-    return ((n % 1) === 0 ? n : Math.floor(n));
-  }
-  return defaultVal;
-}
-
-function writeU_Int8(buf, value, offset, min, max) {
-  value = +value;
-  validateNumber(offset, "offset");
-  if (value > max || value < min) {
-    throw new ERR_OUT_OF_RANGE("value", `>= ${min} and <= ${max}`, value);
-  }
-  if (buf[offset] === undefined) {
-    boundsError(offset, buf.length - 1);
-  }
-
-  buf[offset] = value;
-  return offset + 1;
-}
-
-function writeU_Int16BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 1);
-
-  buf[offset++] = value >>> 8;
-  buf[offset++] = value;
-  return offset;
-}
-
-function _writeUInt32LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 3);
-
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  return offset;
-}
-
-function writeU_Int16LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 1);
-
-  buf[offset++] = value;
-  buf[offset++] = value >>> 8;
-  return offset;
-}
-
-function _writeUInt32BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 3);
-
-  buf[offset + 3] = value;
-  value = value >>> 8;
-  buf[offset + 2] = value;
-  value = value >>> 8;
-  buf[offset + 1] = value;
-  value = value >>> 8;
-  buf[offset] = value;
-  return offset + 4;
-}
-
-function writeU_Int48BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 5);
-
-  const newVal = Math.floor(value * 2 ** -32);
-  buf[offset++] = newVal >>> 8;
-  buf[offset++] = newVal;
-  buf[offset + 3] = value;
-  value = value >>> 8;
-  buf[offset + 2] = value;
-  value = value >>> 8;
-  buf[offset + 1] = value;
-  value = value >>> 8;
-  buf[offset] = value;
-  return offset + 4;
-}
-
-function writeU_Int40BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 4);
-
-  buf[offset++] = Math.floor(value * 2 ** -32);
-  buf[offset + 3] = value;
-  value = value >>> 8;
-  buf[offset + 2] = value;
-  value = value >>> 8;
-  buf[offset + 1] = value;
-  value = value >>> 8;
-  buf[offset] = value;
-  return offset + 4;
-}
-
-function writeU_Int32BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 3);
-
-  buf[offset + 3] = value;
-  value = value >>> 8;
-  buf[offset + 2] = value;
-  value = value >>> 8;
-  buf[offset + 1] = value;
-  value = value >>> 8;
-  buf[offset] = value;
-  return offset + 4;
-}
-
-function writeU_Int24BE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 2);
-
-  buf[offset + 2] = value;
-  value = value >>> 8;
-  buf[offset + 1] = value;
-  value = value >>> 8;
-  buf[offset] = value;
-  return offset + 3;
-}
-
-const validateOffset = (
-  value,
-  name,
-  min = 0,
-  max = Number.MAX_SAFE_INTEGER,
-) => {
-  if (typeof value !== "number") {
-    throw new ERR_INVALID_ARG_TYPE(name, "number", value);
-  }
-  if (!Number.isInteger(value)) {
-    throw new ERR_OUT_OF_RANGE(name, "an integer", value);
-  }
-  if (value < min || value > max) {
-    throw new ERR_OUT_OF_RANGE(name, `>= ${min} && <= ${max}`, value);
-  }
-};
-
-function writeU_Int48LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 5);
-
-  const newVal = Math.floor(value * 2 ** -32);
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  buf[offset++] = newVal;
-  buf[offset++] = newVal >>> 8;
-  return offset;
-}
-
-function writeU_Int40LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 4);
-
-  const newVal = value;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  buf[offset++] = Math.floor(newVal * 2 ** -32);
-  return offset;
-}
-function writeU_Int32LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 3);
-
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  return offset;
-}
-
-function writeUInt32LE(value, offset = 0) {
-  return writeU_Int32LE(this, value, offset, 0, 0xffffffff);
-}
-
-function writeU_Int24LE(buf, value, offset, min, max) {
-  value = +value;
-  checkInt(value, min, max, buf, offset, 2);
-
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  value = value >>> 8;
-  buf[offset++] = value;
-  return offset;
-}
+import {
+  _copyActual,
+  _writeUInt32BE,
+  _writeUInt32LE,
+  base64UrlToBytes,
+  bigEndian,
+  boundsError,
+  byteLengthUtf8,
+  getEncodingOps,
+  isUint8Array,
+  readDoubleBackwards,
+  readDoubleForwards,
+  readFloatBackwards,
+  readFloatForwards,
+  readInt24BE,
+  readInt24LE,
+  readInt40BE,
+  readInt40LE,
+  readInt48BE,
+  readInt48LE,
+  readUInt16BE,
+  readUInt24BE,
+  readUInt24LE,
+  readUInt32BE,
+  readUInt40BE,
+  readUInt40LE,
+  readUInt48BE,
+  readUInt48LE,
+  toInteger,
+  validateNumber,
+  validateOffset,
+  writeDoubleBackwards,
+  writeDoubleForwards,
+  writeFloatBackwards,
+  writeFloatForwards,
+  writeU_Int16BE,
+  writeU_Int16LE,
+  writeU_Int24BE,
+  writeU_Int24LE,
+  writeU_Int32BE,
+  writeU_Int32LE,
+  writeU_Int40BE,
+  writeU_Int40LE,
+  writeU_Int48BE,
+  writeU_Int48LE,
+  writeU_Int8,
+} from "./internal/buffer.js";
 
 // TODO(Soremwar)
 // Find correct crediting and license
@@ -886,103 +195,11 @@ var require_base64_js = __commonJS({
   },
 });
 
-// node_modules/ieee754/index.js
-var require_ieee754 = __commonJS({
-  "node_modules/ieee754/index.js"(exports) {
-    exports.read = function (buffer, offset, isLE, mLen, nBytes) {
-      var e, m;
-      var eLen = nBytes * 8 - mLen - 1;
-      var eMax = (1 << eLen) - 1;
-      var eBias = eMax >> 1;
-      var nBits = -7;
-      var i = isLE ? nBytes - 1 : 0;
-      var d = isLE ? -1 : 1;
-      var s = buffer[offset + i];
-      i += d;
-      e = s & (1 << -nBits) - 1;
-      s >>= -nBits;
-      nBits += eLen;
-      for (; nBits > 0; e = e * 256 + buffer[offset + i], i += d, nBits -= 8) {
-      }
-      m = e & (1 << -nBits) - 1;
-      e >>= -nBits;
-      nBits += mLen;
-      for (; nBits > 0; m = m * 256 + buffer[offset + i], i += d, nBits -= 8) {
-      }
-      if (e === 0) {
-        e = 1 - eBias;
-      } else if (e === eMax) {
-        return m ? NaN : (s ? -1 : 1) * Infinity;
-      } else {
-        m = m + Math.pow(2, mLen);
-        e = e - eBias;
-      }
-      return (s ? -1 : 1) * m * Math.pow(2, e - mLen);
-    };
-    exports.write = function (buffer, value, offset, isLE, mLen, nBytes) {
-      var e, m, c;
-      var eLen = nBytes * 8 - mLen - 1;
-      var eMax = (1 << eLen) - 1;
-      var eBias = eMax >> 1;
-      var rt = mLen === 23 ? Math.pow(2, -24) - Math.pow(2, -77) : 0;
-      var i = isLE ? 0 : nBytes - 1;
-      var d = isLE ? 1 : -1;
-      var s = value < 0 || value === 0 && 1 / value < 0 ? 1 : 0;
-      value = Math.abs(value);
-      if (isNaN(value) || value === Infinity) {
-        m = isNaN(value) ? 1 : 0;
-        e = eMax;
-      } else {
-        e = Math.floor(Math.log(value) / Math.LN2);
-        if (value * (c = Math.pow(2, -e)) < 1) {
-          e--;
-          c *= 2;
-        }
-        if (e + eBias >= 1) {
-          value += rt / c;
-        } else {
-          value += rt * Math.pow(2, 1 - eBias);
-        }
-        if (value * c >= 2) {
-          e++;
-          c /= 2;
-        }
-        if (e + eBias >= eMax) {
-          m = 0;
-          e = eMax;
-        } else if (e + eBias >= 1) {
-          m = (value * c - 1) * Math.pow(2, mLen);
-          e = e + eBias;
-        } else {
-          m = value * Math.pow(2, eBias - 1) * Math.pow(2, mLen);
-          e = 0;
-        }
-      }
-      for (
-        ;
-        mLen >= 8;
-        buffer[offset + i] = m & 255, i += d, m /= 256, mLen -= 8
-      ) {
-      }
-      e = e << mLen | m;
-      eLen += mLen;
-      for (
-        ;
-        eLen > 0;
-        buffer[offset + i] = e & 255, i += d, e /= 256, eLen -= 8
-      ) {
-      }
-      buffer[offset + i - d] |= s * 128;
-    };
-  },
-});
-
 // node_modules/buffer/index.js
 var require_buffer = __commonJS({
   "node_modules/buffer/index.js"(exports) {
     "use strict";
     var base64 = require_base64_js();
-    var ieee754 = require_ieee754();
     var customInspectSymbol =
       typeof Symbol === "function" && typeof Symbol["for"] === "function"
         ? Symbol["for"]("nodejs.util.inspect.custom")
